Fontifies function declarations
This commit adds a function to match function declarations.

We search forward for a lower camel case identifier followed by an
opening paren, then jump to closing paren, and check for either an async
keyword, or an opening of a block statement, or an arrow.

The docstring of `font-lock-keywords` specifies how a matcher function
must work.

Function takes an argument `limit`, searches from point, and if no
match found returns nil, otherwise, sets match data, sets point for
next search, and returns non-nil.