Add fontification tests
Use faceup package to test fontification.

The test dart-single-quoted-string-test has a conditional which shows
single quoted strings are failing starting at emacs 26.

The setup function dart-font-lock-test has a let for
c-standard-font-lock-fontify-region-function so that it's not nil, as
that was breaking the tests on emacs 24.3.

We also let dart-enable-analysis-server to nil, since there is a
separate test for that, that is, regarding
dart-add-analysis-root-for-file.